% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions for RSA and CCA.R
\name{adjust_CCA}
\alias{adjust_CCA}
\title{Adjusted Canonical Correlation Analysis}
\usage{
adjust_CCA(X, Y, num_permutations = 1000, raw_CCA)
}
\arguments{
\item{X}{matrix or dataframe, a set of explanatory variables.}

\item{Y}{matrix or dataframe, a set of response variables.}

\item{num_permutations}{numeric, the number of permutations to perform. Default is 1000.}

\item{raw_CCA}{numeric, the observed canonical correlation.}
}
\value{
numeric, the adjusted canonical correlation.
}
\description{
This function calculates an adjusted canonical correlation by comparing the observed canonical correlation
to a null distribution generated by permuting the response variables.
}
\examples{
# Example with random data
set.seed(1)
X <- matrix(rnorm(100*3), 100, 3)
Y <- matrix(rnorm(100*2), 100, 2)
raw_CCA <- max(stats::cancor(X, Y)$cor)
adjusted_CCA <- adjust_CCA(X, Y, num_permutations = 500, raw_CCA = raw_CCA)
}
