% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions for RSA and CCA.R
\name{simulate_heterogeneous_data}
\alias{simulate_heterogeneous_data}
\title{Simulate Heterogeneous Data}
\usage{
simulate_heterogeneous_data(n, means, cor_values)
}
\arguments{
\item{n}{numeric, a single number specifying the sample size.}

\item{means}{numeric, a vector specifying the means for each variable.}

\item{cor_values}{numeric, a vector specifying the unique values in the lower
triangle of the correlation matrix, excluding the diagonal}
}
\value{
A matrix containing the multivariate dataset
}
\description{
The function creates multivariate data consisting of heterogeneous subsets.
}
\examples{
# Create a matrix with 4 variables and 100 subjects
n <- 100
means <- rep(0, 4)
cor_values <- c(0.1, 0.2, 0.3, 0.4, 0.5, 0.6)
my_data <- simulate_multivariate_data(n, means, cor_values)
}
